Copy scene elements to context before rendering
Take a snapshot of each scene element as a context element.

This should allow manipulating scene elements during event dispatch without causing problems with the renderer if it is currently rendering the scene.

Example: changing veretex buffer on a geometry, the context holds a copy so the old buffers aren't freed until the frame is finished.